ensure reactive array watchers can be blocked internaly to prevel update loops

+++ src/js/modules/reactive_data.js
@@ -17,36 +17,36 @@ ReactiveData.prototype.watchData = function(data){
 	self.origFuncs.push = data.push;
 
 	Object.defineProperty(self.data, "push", {
-	  enumerable: false,
-	  configurable: false, // prevent further meddling...
-	  writable: false, // see above ^
-	  value: function () {
-	  	var args = Array.from(arguments);
+		enumerable: false,
+		value: function () {
+			var args = Array.from(arguments);
 
-	  	args.forEach(function (arg){
-	  		self.table.rowManager.addRowActual(arg, false);
-	  	});
+			if(!self.blocked){
+				args.forEach(function (arg){
+					self.table.rowManager.addRowActual(arg, false);
+				});
+			}
 
-	    return self.origFuncs.push.apply(data, arguments);
-	  }
+			return self.origFuncs.push.apply(data, arguments);
+		}
 	});
 
 	//override array unshift function
 	self.origFuncs.unshift = data.unshift;
 
 	Object.defineProperty(self.data, "unshift", {
-	  enumerable: false,
-	  configurable: false, // prevent further meddling...
-	  writable: false, // see above ^
-	  value: function () {
-	  	var args = Array.from(arguments);
+		enumerable: false,
+		value: function () {
+			var args = Array.from(arguments);
 
-	  	args.forEach(function (arg){
-	  		self.table.rowManager.addRowActual(arg, true);
-	  	});
+			if(!self.blocked){
+				args.forEach(function (arg){
+					self.table.rowManager.addRowActual(arg, true);
+				});
+			}
 
-	    return self.origFuncs.unshift.apply(data, arguments);
-	  }
+			return self.origFuncs.unshift.apply(data, arguments);
+		}
 	});
 
 
@@ -54,44 +54,42 @@ ReactiveData.prototype.watchData = function(data){
 	self.origFuncs.shift = data.shift;
 
 	Object.defineProperty(self.data, "shift", {
-	  enumerable: false,
-	  configurable: false, // prevent further meddling...
-	  writable: false, // see above ^
-	  value: function () {
-	  	var row;
+		enumerable: false,
+		value: function () {
+			var row;
 
-		  if(self.data.length){
-		  	row = self.table.rowManager.getRowFromDataObject(self.data[0]);
+			if(!self.blocked){
+				if(self.data.length){
+					row = self.table.rowManager.getRowFromDataObject(self.data[0]);
 
-		  	if(row){
-		  		row.deleteActual();
-		  	}
-		  }
+					if(row){
+						row.deleteActual();
+					}
+				}
+			}
 
-	    return self.origFuncs.shift.call(data);
-	  }
+			return self.origFuncs.shift.call(data);
+		}
 	});
 
 	//override array pop function
 	self.origFuncs.pop = data.pop;
 
 	Object.defineProperty(self.data, "pop", {
-	  enumerable: false,
-	  configurable: false, // prevent further meddling...
-	  writable: false, // see above ^
-	  value: function () {
-	  	var row;
+		enumerable: false,
+		value: function () {
+			var row;
+			if(!self.blocked){
+				if(self.data.length){
+					row = self.table.rowManager.getRowFromDataObject(self.data[self.data.length - 1]);
 
-		  if(self.data.length){
-		  	row = self.table.rowManager.getRowFromDataObject(self.data[self.data.length - 1]);
-
-		  	if(row){
-		  		row.deleteActual();
-		  	}
-		  }
-
-	    return self.origFuncs.pop.call(data);
-	  }
+					if(row){
+						row.deleteActual();
+					}
+				}
+			}
+			return self.origFuncs.pop.call(data);
+		}
 	});
 };
 
@@ -99,8 +97,8 @@ ReactiveData.prototype.unwatchData = function(){
 	if(this.data !== false){
 		for(var key in this.origFuncs){
 			Object.defineProperty(self.data, key, {
-			  enumerable: false,
-			  value: origFuncs.push,
+				enumerable: false,
+				value: origFuncs.push,
 			});
 		}
 	}
